feat: add `Evm` and `Env` to the API (+snailtracer test) (#196)
